refactor(inji-245): modify the credential endpoint and aud from config
